requirements.txt
tk
PySocks